+++ platform/linux/window-x11.c
@@ -39,6 +39,18 @@
 
 #define MODULE_NAME "window-x11"
 
+static void init_info(struct p_window_info *info_p,
+    const xcb_screen_t *screen, const rect_t *area, const u32 flags);
+static xcb_window_t create_window(
+    const xcb_screen_t *screen, const struct p_window_info *info,
+    xcb_connection_t *conn, const struct libxcb *xcb
+);
+static i32 set_window_properties(struct window_x11_atoms *atoms,
+    const struct p_window_info *info, const char *title, xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb);
+static i32 init_xi2_input(struct window_x11_input *i, xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb);
+
 static i32 init_acceleration(struct window_x11 *win, enum p_window_flags flags);
 
 static i32 render_init_egl(struct x11_render_egl_ctx *egl_rctx,
@@ -46,24 +58,23 @@ static i32 render_init_egl(struct x11_render_egl_ctx *egl_rctx,
 static void render_destroy_egl(struct x11_render_egl_ctx *egl_rctx,
     const struct libxcb *xcb);
 
-static i32 intern_atom(struct window_x11 *win,
-    const char *atom_name, xcb_atom_t *o);
+static i32 intern_atom(const char *atom_name, xcb_atom_t *o,
+    xcb_connection_t *conn, const struct libxcb *xcb);
 static bool query_extension(const char *name,
     xcb_connection_t *conn, const struct libxcb *xcb);
 
 static i32 get_master_input_devices(
-    struct window_x11 *win,
     xcb_input_device_id_t *master_mouse_id,
-    xcb_input_device_id_t *master_keyboard_id
+    xcb_input_device_id_t *master_keyboard_id,
+    xcb_connection_t *conn, const struct libxcb *xcb
 );
 
-static void send_dummy_event_to_self(struct window_x11 *win);
+static void send_dummy_event_to_self(xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb);
 
 i32 window_X11_open(struct window_x11 *win, struct p_window_info *info,
     const char *title, const rect_t *area, const u32 flags)
 {
-#define libxcb_error() (e = win->xcb.xcb_request_check(win->conn, vc), e != NULL)
-
     /* Used for error checking */
     xcb_generic_error_t *e = NULL;
     xcb_void_cookie_t vc = { 0 };
@@ -72,6 +83,8 @@ i32 window_X11_open(struct window_x11 *win, struct p_window_info *info,
     memset(win, 0, sizeof(struct window_x11));
     win->exists = true;
 
+    /* Check the parameters for compatibility with the 1980s-like
+     * X11 protocol constraints */
     if (area->w > UINT16_MAX || area->h > UINT16_MAX
         || area->x < INT16_MIN || area->x > INT16_MAX
         || area->y < INT16_MIN || area->y > INT16_MAX)
@@ -83,193 +96,74 @@ i32 window_X11_open(struct window_x11 *win, struct p_window_info *info,
             INT16_MIN, INT16_MAX, 0, UINT16_MAX);
     }
 
-    win->generic_info_p = info;
+    /* Load the requried libraries (`win->xcb`) */
+    if (libxcb_load(&win->xcb))
+        goto_error("Failed to load libxcb");
 
-    win->registered_keyboard = NULL;
-    win->keyboard_deregistration_ack = p_mt_cond_create();
-    atomic_store(&win->keyboard_deregistration_notify, false);
-
-    win->registered_mouse = NULL;
-    win->mouse_deregistration_ack = p_mt_cond_create();
-    atomic_store(&win->mouse_deregistration_notify, false);
-
-    if (libxcb_load(&win->xcb)) goto_error("Failed to load libxcb");
-
-    /* Open a connection */
+    /* Open a connection (`win->conn`) */
     win->conn = win->xcb.xcb_connect(NULL, NULL);
     if (win->xcb.xcb_connection_has_error(win->conn))
         goto_error("Failed to connect to the X server");
 
-    /* Initialize Xinput v2 */
-    if (X11_check_xinput2_extension(win->conn, &win->xcb))
-        goto_error("The XInput2 extension is not available");
+    /* Note that this macro should only be used right after calling
+     * a `*_checked` xcb function. Otherwise it's useless */
+#define libxcb_error() \
+    (e = win->xcb.xcb_request_check(win->conn, vc), e != NULL)
 
-    win->xinput_ext_data = win->xcb.xcb_get_extension_data(win->conn,
-        win->xcb.xinput.xcb_input_id);
-    if (win->xinput_ext_data == NULL)
-        goto_error("Couldn't get XInput extension data");
+    /* Get the server screen configuration (`win->screen`) */
 
-    /* Get the X server setup */
-    win->setup = win->xcb.xcb_get_setup(win->conn);
-    if (win->setup == NULL) goto_error("Failed to get X setup");
+    /* We need to get the screen info early because
+     * `init_info` needs it */
+    const struct xcb_setup_t *setup = win->xcb.xcb_get_setup(win->conn);
+    if (setup == NULL)
+        goto_error("Failed to get X server setup");
 
     /* Get the current screen */
-    win->iter = win->xcb.xcb_setup_roots_iterator(win->setup);
-    win->screen = win->iter.data;
+    xcb_screen_iterator_t scr_iter = win->xcb.xcb_setup_roots_iterator(setup);
+    win->screen = scr_iter.data;
 
-    /* Generate the window ID */
-    win->win = win->xcb.xcb_generate_id(win->conn);
+    /* Initialize the generic window info struct
+     * (`win->generic_info_p`) */
+    win->generic_info_p = info;
+    init_info(win->generic_info_p, win->screen, area, flags);
 
-    /* Handle WINDOW_POS_CENTERED flags */
-    i16 x = (i16)area->x, y = (i16)area->y;
-    if (flags & P_WINDOW_POS_CENTERED_X)
-        x = (win->screen->width_in_pixels - area->w) / 2;
-    if (flags & P_WINDOW_POS_CENTERED_Y)
-        y = (win->screen->height_in_pixels - area->h) / 2;
-
-    /* Initialize the generic window parameters */
-    info->client_area.x = (i32)x;
-    info->client_area.y = (i32)y;
-    info->client_area.w = area->w;
-    info->client_area.h = area->h;
-
-    info->display_rect.x = info->display_rect.y = 0;
-    info->display_rect.w = win->screen->width_in_pixels;
-    info->display_rect.h = win->screen->height_in_pixels;
-    info->display_color_format = BGRX32;
-
-    info->gpu_acceleration = P_WINDOW_ACCELERATION_UNSET_;
-    info->vsync_supported = false;
-
-    /* Create the window */
-    u32 value_mask = XCB_CW_BACK_PIXEL | XCB_CW_EVENT_MASK;
-    u32 value_list[] = {
-        win->screen->black_pixel,
-        XCB_EVENT_MASK_STRUCTURE_NOTIFY | XCB_EVENT_MASK_EXPOSURE
-    };
-
-    vc = win->xcb.xcb_create_window_checked(win->conn, XCB_COPY_FROM_PARENT,
-        win->win, win->screen->root,
-        x, y, area->w, area->h, 0, XCB_WINDOW_CLASS_INPUT_OUTPUT,
-        win->screen->root_visual, value_mask, value_list);
-    if (libxcb_error()) goto_error("Failed to create the window");
-    win->win_created = true;
-
-    /* Get the UTF8 string atom */
-    if (intern_atom(win, "UTF8_STRING", &win->UTF8_STRING)) goto err;
-
-    /* Set the window title */
-    vc = win->xcb.xcb_change_property_checked(win->conn, XCB_PROP_MODE_REPLACE,
-        win->win, XCB_ATOM_WM_NAME, XCB_ATOM_STRING, 8,
-        strlen((char *)title), title
-    );
-    if (libxcb_error()) goto_error("Failed to change window name");
-
-    if (intern_atom(win, "_NET_WM_NAME", &win->NET_WM_NAME)) goto err;
-    vc = win->xcb.xcb_change_property_checked(win->conn, XCB_PROP_MODE_REPLACE,
-        win->win, win->NET_WM_NAME, win->UTF8_STRING, 8,
-        strlen((char *)title), title
-    );
-    if (libxcb_error()) goto_error("Failed to set the _NET_WM_NAME property");
-
-    /* Set the window to floating */
-    if (intern_atom(win,
-            "_NET_WM_STATE_ABOVE", &win->NET_WM_STATE_ABOVE
-        )
-    ) goto err;
-
-    i32 net_wm_state_above_val = 1;
-    vc = win->xcb.xcb_change_property_checked(win->conn, XCB_PROP_MODE_REPLACE,
-        win->win, win->NET_WM_STATE_ABOVE, XCB_ATOM_INTEGER, 32,
-        1, &net_wm_state_above_val
-    );
-    if (libxcb_error())
-        goto_error("Failed to set the NET_WM_STATE_ABOVE property");
-
-    /* Set the window minimum and maximum size */
-    xcb_size_hints_t hints = { 0 };
-    win->xcb.xcb_icccm_size_hints_set_min_size(&hints, area->w, area->h);
-    win->xcb.xcb_icccm_size_hints_set_max_size(&hints, area->w, area->h);
-    vc = win->xcb.xcb_icccm_set_wm_normal_hints_checked(win->conn,
-            win->win, &hints);
-    if (libxcb_error()) goto_error("Failed to set WM normal hints");
-
-
-    /* Set the WM_DELETE_WINDOW protocol atom */
-    if (intern_atom(win, "WM_PROTOCOLS", &win->WM_PROTOCOLS))
-        goto err;
-    if (intern_atom(win, "WM_DELETE_WINDOW", &win->WM_DELETE_WINDOW))
+    /* Create the window (`win->win_handle`) */
+    win->win_handle = create_window(win->screen, win->generic_info_p,
+        win->conn, &win->xcb);
+    if (win->win_handle == XCB_NONE)
         goto err;
 
-    vc = win->xcb.xcb_change_property_checked(win->conn, XCB_PROP_MODE_REPLACE,
-        win->win, win->WM_PROTOCOLS, XCB_ATOM_ATOM, 32,
-        1, &win->WM_DELETE_WINDOW
-    );
-    if (libxcb_error()) goto_error("Failed to set WM protocols");
+    /* Set window properties such as the title, floating state,
+     * minimal and maximal size hints, etc
+     * (`win->atoms`) */
+    if (set_window_properties(&win->atoms, win->generic_info_p, title,
+            win->win_handle, win->conn, &win->xcb)
+    ) {
+        goto_error("Failed to set window properties");
+    }
 
-    /* Initialize the graphics context */
-
-    /* Initialize the XInput2 externsion */
-    /* Get the master keyboard and master mouse device IDs */
-    if (get_master_input_devices(win,
-            &win->master_mouse_id, &win->master_keyboard_id
-        )
-    ) goto_error("Failed to query master input device IDs");
-
-    /* Credits: https://gist.github.com/LemonBoy/dfe1d7ea428794c65b3d
-     * Like come on xcb, WTF?!??!?!! */
-    const struct xi2_mask {
-        const xcb_input_event_mask_t head;
-        const xcb_input_xi_event_mask_t mask;
-    } keyboard_mask = {
-        .head = {
-            .deviceid = win->master_keyboard_id,
-            .mask_len = sizeof(keyboard_mask.mask) / sizeof(u32),
-        },
-        .mask = XCB_INPUT_XI_EVENT_MASK_KEY_PRESS
-        | XCB_INPUT_XI_EVENT_MASK_KEY_RELEASE,
-    }, mouse_mask = {
-        .head = {
-            .deviceid = win->master_mouse_id,
-            .mask_len = sizeof(mouse_mask.mask) / sizeof(u32),
-        },
-        .mask = XCB_INPUT_XI_EVENT_MASK_MOTION
-        | XCB_INPUT_XI_EVENT_MASK_BUTTON_PRESS
-        | XCB_INPUT_XI_EVENT_MASK_BUTTON_RELEASE,
-    };
-
-    vc = win->xcb.xinput.xcb_input_xi_select_events_checked(
-        win->conn, win->win, 1, &keyboard_mask.head
-    );
+    /* Map the window to make it visible */
+    vc = win->xcb.xcb_map_window_checked(win->conn, win->win_handle);
     if (libxcb_error())
-        goto_error("Failed to enable keyboard input handling with Xi2");
+        goto_error("Failed to map (show) the window");
 
+    /* Initialize input stuff (`win->input`) */
+    if (init_xi2_input(&win->input, win->win_handle, win->conn, &win->xcb))
+        goto_error("Failed to initialize input");
 
-    vc = win->xcb.xinput.xcb_input_xi_select_events_checked(
-        win->conn, win->win, 1, &mouse_mask.head
-    );
-    if (libxcb_error())
-        goto_error("Failed to enable mouse input handling with Xi2");
-
-    /* Allocate the key symbols struct, used by keyboard-x11
-     * to map keycodes received from events to keysyms */
-    win->key_symbols = win->xcb.xcb_key_symbols_alloc(win->conn);
-    if (win->key_symbols == NULL)
-        goto_error("Failed to allocate key symbols");
-
-    /* Initialize the GPU acceleration based on the flags */
+    /* Initialize the GPU acceleration based on the flags.
+     * This populates `win->render` as well as
+     *  the `gpu_acceleration` and `vsync_supported` fields in
+     * `win->generic_info_p`. */
     if (init_acceleration(win, flags))
         goto_error("Failed to initialize GPU acceleration");
 
-    /* Map the window so that it's visible */
-    vc = win->xcb.xcb_map_window_checked(win->conn, win->win);
-    if (libxcb_error()) goto_error("Failed to map (show) the window");
-
-    /* Finally, flush all the commands */
+    /* Flush all the commands that are still queued */
     if (win->xcb.xcb_flush(win->conn) <= 0)
         goto_error("Failed to flush xcb requests");
 
-    /* Create the thread that listens for events */
+    /* Create the thread that listens for events
+     * (`win->listener`) */
     atomic_init(&win->listener.running, true);
     if (p_mt_thread_create(&win->listener.thread,
         window_X11_event_listener_fn, win))
@@ -277,9 +171,11 @@ i32 window_X11_open(struct window_x11 *win, struct p_window_info *info,
         goto_error("Failed to create listener thread.");
     }
 
-    s_log_debug("%s() OK; Screen is %ux%u, use shm: %i", __func__,
+    s_log_debug("%s() OK; Screen is %ux%u, use shm: %d, use vsync: %d",
+        __func__,
         win->screen->width_in_pixels, win->screen->height_in_pixels,
-        win->xcb.shm.loaded_);
+        win->xcb.shm.loaded_, win->generic_info_p->vsync_supported
+    );
     return 0;
 
 err:
@@ -294,22 +190,18 @@ err:
 void window_X11_close(struct window_x11 *win)
 {
     s_assert(win->exists, "Attempt to double-free window");
+    win->exists = false;
 
     s_log_verbose("Destroying X11 window...");
     if (win->xcb.loaded_) {
-        /* Free acceleration-specific resources */
-        if (win->conn)
-            window_X11_set_acceleration(win, P_WINDOW_ACCELERATION_UNSET_);
-
-        if (win->key_symbols) win->xcb.xcb_key_symbols_free(win->key_symbols);
-
+        /* End the listener thread (`win->listener`) */
         if (atomic_load(&win->listener.running)) {
             atomic_store(&win->listener.running, false);
 
-            if (win->win_created) {
+            if (win->win_handle != XCB_NONE) {
                 /* Send an event to our window to interrupt the blocking
                  * `xcb_wait_for_event` call in the listener thread */
-                send_dummy_event_to_self(win);
+                send_dummy_event_to_self(win->win_handle, win->conn, &win->xcb);
                 p_mt_thread_wait(&win->listener.thread);
             } else {
                 /* Forcibly terminate the listener if (somehow)
@@ -318,16 +210,44 @@ void window_X11_close(struct window_x11 *win)
             }
         }
 
-        if (win->win_created) win->xcb.xcb_destroy_window(win->conn, win->win);
-        if (win->conn) win->xcb.xcb_disconnect(win->conn);
+        /* Free acceleration-specific resources
+         * (`win->render`) */
+        if (win->conn)
+            window_X11_set_acceleration(win, P_WINDOW_ACCELERATION_UNSET_);
+
+        /* Destroy anything related to input
+         * (`win->input`) */
+        if (win->input.key_symbols)
+            win->xcb.xcb_key_symbols_free(win->input.key_symbols);
+
+        p_mt_cond_destroy(&win->input.mouse_deregistration_ack);
+        p_mt_cond_destroy(&win->input.keyboard_deregistration_ack);
+
+        /* Reset the atoms (`win->atoms`) */
+        memset(&win->atoms, 0, sizeof(struct window_x11_atoms));
+
+        /* Destroy the window itself (`win->win_handle`) */
+        if (win->win_handle != XCB_NONE) {
+            win->xcb.xcb_destroy_window(win->conn, win->win_handle);
+            win->win_handle = XCB_NONE;
+        }
+
+        /* Reset the info struct pointer (`win->generic_info_p`) */
+        win->generic_info_p = NULL;
+
+        /* Delete the screen info pointer (`win->screen`) */
+        win->screen = NULL;
+
+        /* Finally, close the connection and free any xcb resources
+         * (`win->conn`) */
+        if (win->conn) {
+            win->xcb.xcb_disconnect(win->conn);
+            win->conn = NULL;
+        }
     }
+
+    /* Unload the xcb libraries (`win->xcb`) */
     libxcb_unload(&win->xcb);
-
-    p_mt_cond_destroy(&win->mouse_deregistration_ack);
-    p_mt_cond_destroy(&win->keyboard_deregistration_ack);
-
-    /* win->exists is also set to false */
-    memset(win, 0, sizeof(struct window_x11));
 }
 
 struct pixel_flat_data * window_X11_swap_buffers(struct window_x11 *win,
@@ -346,7 +266,7 @@ struct pixel_flat_data * window_X11_swap_buffers(struct window_x11 *win,
     }
 
     return X11_render_present_software(&win->render.sw,
-        win->win, win->conn, &win->xcb, present_mode);
+        win->win_handle, win->conn, &win->xcb, present_mode);
 }
 
 i32 window_X11_register_keyboard(struct window_x11 *win,
@@ -354,9 +274,9 @@ i32 window_X11_register_keyboard(struct window_x11 *win,
 {
     u_check_params(win != NULL && kb != NULL);
 
-    if (win->registered_keyboard != NULL) return 1;
+    if (win->input.registered_keyboard != NULL) return 1;
 
-    win->registered_keyboard = kb;
+    win->input.registered_keyboard = kb;
     return 0;
 }
 
@@ -364,9 +284,9 @@ i32 window_X11_register_mouse(struct window_x11 *win, struct mouse_x11 *mouse)
 {
     u_check_params(win != NULL && mouse != NULL);
 
-    if (win->registered_mouse != NULL) return 1;
+    if (win->input.registered_mouse != NULL) return 1;
 
-    win->registered_mouse = mouse;
+    win->input.registered_mouse = mouse;
     return 0;
 }
 
@@ -374,18 +294,18 @@ void window_X11_deregister_keyboard(struct window_x11 *win)
 {
     u_check_params(win != NULL);
 
-    if (win->registered_keyboard != NULL) {
+    if (win->input.registered_keyboard != NULL) {
         /* Notify the event thread that the keyboard is
          * being deregistered and wait for it to acknowledge that */
         p_mt_mutex_t tmp_mutex = p_mt_mutex_create();
         p_mt_mutex_lock(&tmp_mutex);
 
-        atomic_store(&win->keyboard_deregistration_notify, true);
-        send_dummy_event_to_self(win);
-        p_mt_cond_wait(win->keyboard_deregistration_ack, tmp_mutex);
+        atomic_store(&win->input.keyboard_deregistration_notify, true);
+        send_dummy_event_to_self(win->win_handle, win->conn, &win->xcb);
+        p_mt_cond_wait(win->input.keyboard_deregistration_ack, tmp_mutex);
 
-        win->registered_keyboard = NULL;
-        atomic_store(&win->keyboard_deregistration_notify, false);
+        win->input.registered_keyboard = NULL;
+        atomic_store(&win->input.keyboard_deregistration_notify, false);
 
         p_mt_mutex_destroy(&tmp_mutex);
     }
@@ -394,18 +314,18 @@ void window_X11_deregister_keyboard(struct window_x11 *win)
 void window_X11_deregister_mouse(struct window_x11 *win)
 {
     u_check_params(win != NULL);
-    if (win->registered_mouse != NULL) {
+    if (win->input.registered_mouse != NULL) {
         /* Notify the event thread that the mouse is
          * being deregistered and wait for it to acknowledge that */
         p_mt_mutex_t tmp_mutex = p_mt_mutex_create();
         p_mt_mutex_lock(&tmp_mutex);
 
-        atomic_store(&win->mouse_deregistration_notify, true);
-        send_dummy_event_to_self(win);
-        p_mt_cond_wait(win->mouse_deregistration_ack, tmp_mutex);
+        atomic_store(&win->input.mouse_deregistration_notify, true);
+        send_dummy_event_to_self(win->win_handle, win->conn, &win->xcb);
+        p_mt_cond_wait(win->input.mouse_deregistration_ack, tmp_mutex);
 
-        win->registered_mouse = NULL;
-        atomic_store(&win->mouse_deregistration_notify, false);
+        win->input.registered_mouse = NULL;
+        atomic_store(&win->input.mouse_deregistration_notify, false);
 
         p_mt_mutex_destroy(&tmp_mutex);
     }
@@ -427,7 +347,7 @@ i32 window_X11_set_acceleration(struct window_x11 *win,
     switch (old_acceleration) {
         case P_WINDOW_ACCELERATION_NONE:
             X11_render_destroy_software(&win->render.sw,
-                win->conn, win->win, &win->xcb);
+                win->conn, win->win_handle, &win->xcb);
             break;
         case P_WINDOW_ACCELERATION_OPENGL:
             render_destroy_egl(&win->render.egl, &win->xcb);
@@ -447,7 +367,7 @@ i32 window_X11_set_acceleration(struct window_x11 *win,
         if (X11_render_init_software(&win->render.sw,
                 win->generic_info_p->client_area.w,
                 win->generic_info_p->client_area.h,
-                win->screen->root_depth, win->win, win->conn, &win->xcb,
+                win->screen->root_depth, win->win_handle, win->conn, &win->xcb,
                 &win->generic_info_p->vsync_supported))
         {
             s_log_error("Failed to set up the window for software rendering.");
@@ -608,6 +528,231 @@ i32 X11_check_present_extension(xcb_connection_t *conn,
     return 0;
 }
 
+static void init_info(struct p_window_info *info_p,
+    const xcb_screen_t *screen, const rect_t *area, const u32 flags)
+{
+    /* Handle WINDOW_POS_CENTERED flags */
+    i16 x = (i16)area->x, y = (i16)area->y;
+    if (flags & P_WINDOW_POS_CENTERED_X)
+        x = (screen->width_in_pixels - area->w) / 2;
+    if (flags & P_WINDOW_POS_CENTERED_Y)
+        y = (screen->height_in_pixels - area->h) / 2;
+
+    info_p->client_area.x = (i32)x;
+    info_p->client_area.y = (i32)y;
+    info_p->client_area.w = area->w;
+    info_p->client_area.h = area->h;
+
+    info_p->display_rect.x = info_p->display_rect.y = 0;
+    info_p->display_rect.w = screen->width_in_pixels;
+    info_p->display_rect.h = screen->height_in_pixels;
+    info_p->display_color_format = BGRX32;
+
+    info_p->gpu_acceleration = P_WINDOW_ACCELERATION_UNSET_;
+    info_p->vsync_supported = false; /* To be decided later */
+}
+
+static xcb_window_t create_window(
+    const xcb_screen_t *screen, const struct p_window_info *info,
+    xcb_connection_t *conn, const struct libxcb *xcb
+)
+{
+    /* Generate the window ID */
+    xcb_window_t tmp_window_id = xcb->xcb_generate_id(conn);
+
+    /* Create the window */
+    const u8 DEPTH = XCB_COPY_FROM_PARENT;
+    const xcb_window_t NEW_WINDOW_ID = tmp_window_id;
+    const xcb_window_t PARENT_WINDOW_ID = screen->root;
+    const i16 X = info->client_area.x, Y = info->client_area.y;
+    const u16 W = info->client_area.w, H = info->client_area.h;
+    const u16 BORDER_WIDTH = 0;
+    const u16 WINDOW_CLASS = XCB_WINDOW_CLASS_INPUT_OUTPUT;
+    const xcb_visualid_t VISUAL = screen->root_visual;
+    const u32 VALUE_MASK = XCB_CW_BACK_PIXEL | XCB_CW_EVENT_MASK;
+    const u32 VALUE_LIST[] = {
+        screen->black_pixel,
+        XCB_EVENT_MASK_STRUCTURE_NOTIFY | XCB_EVENT_MASK_EXPOSURE
+    };
+    xcb_void_cookie_t vc = xcb->xcb_create_window_checked(conn,
+        DEPTH, NEW_WINDOW_ID, PARENT_WINDOW_ID, X, Y, W, H,
+        BORDER_WIDTH, WINDOW_CLASS, VISUAL, VALUE_MASK, VALUE_LIST);
+    xcb_generic_error_t *e = NULL;
+    if (e = xcb->xcb_request_check(conn, vc), e != NULL) {
+        s_log_error("Failed to create the window!");
+        u_nfree(&e);
+        return XCB_NONE;
+    }
+
+    return tmp_window_id;
+}
+
+static i32 set_window_properties(struct window_x11_atoms *atoms,
+    const struct p_window_info *info, const char *title, xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb)
+{
+    xcb_void_cookie_t vc;
+    xcb_generic_error_t *e = NULL;
+
+#define libxcb_error() \
+    (e = xcb->xcb_request_check(conn, vc), e != NULL)
+
+    /* Get the UTF8 string atom */
+    if (intern_atom("UTF8_STRING", &atoms->UTF8_STRING, conn, xcb))
+        goto err;
+
+    /* Set the window title */
+    vc = xcb->xcb_change_property_checked(conn, XCB_PROP_MODE_REPLACE,
+        win, XCB_ATOM_WM_NAME, XCB_ATOM_STRING, 8, strlen(title), title);
+    if (libxcb_error())
+        goto_error("Failed to change window name");
+
+    if (intern_atom("_NET_WM_NAME", &atoms->NET_WM_NAME, conn, xcb))
+        goto err;
+    vc = xcb->xcb_change_property_checked(conn, XCB_PROP_MODE_REPLACE,
+        win, atoms->NET_WM_NAME, atoms->UTF8_STRING, 8,
+        strlen((char *)title), title
+    );
+    if (libxcb_error())
+        goto_error("Failed to set the _NET_WM_NAME property");
+
+    /* Set the window to floating */
+    if (intern_atom("_NET_WM_STATE_ABOVE", &atoms->NET_WM_STATE_ABOVE,
+            conn, xcb)
+    ) {
+        goto err;
+    }
+
+    i32 net_wm_state_above_val = 1;
+    vc = xcb->xcb_change_property_checked(conn, XCB_PROP_MODE_REPLACE,
+        win, atoms->NET_WM_STATE_ABOVE, XCB_ATOM_INTEGER, 32,
+        1, &net_wm_state_above_val
+    );
+    if (libxcb_error())
+        goto_error("Failed to set the NET_WM_STATE_ABOVE "
+            "(floating window) property");
+
+    /* Set the window minimum and maximum size to the same value
+     * to mark the window as non-resizable for the window manager */
+    xcb_size_hints_t hints = { 0 };
+    xcb->xcb_icccm_size_hints_set_min_size(&hints,
+        info->client_area.w, info->client_area.h);
+    xcb->xcb_icccm_size_hints_set_max_size(&hints,
+        info->client_area.w, info->client_area.h);
+    vc = xcb->xcb_icccm_set_wm_normal_hints_checked(conn, win, &hints);
+    if (libxcb_error())
+        goto_error("Failed to set WM normal hints");
+
+    /* Set the WM_DELETE_WINDOW protocol atom,
+     * to enable receiving window close events */
+    if (intern_atom("WM_PROTOCOLS", &atoms->WM_PROTOCOLS, conn, xcb))
+        goto err;
+    if (intern_atom("WM_DELETE_WINDOW", &atoms->WM_DELETE_WINDOW,
+            conn, xcb))
+        goto err;
+
+    vc = xcb->xcb_change_property_checked(conn, XCB_PROP_MODE_REPLACE,
+        win, atoms->WM_PROTOCOLS, XCB_ATOM_ATOM, 32,
+        1, &atoms->WM_DELETE_WINDOW
+    );
+    if (libxcb_error())
+        goto_error("Failed to set WM protocols");
+
+    return 0;
+
+err:
+    if (e != NULL) {
+        u_nfree(&e);
+    }
+    return 1;
+
+#undef libxcb_error
+}
+
+static i32 init_xi2_input(struct window_x11_input *i, xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb)
+{
+    xcb_void_cookie_t vc;
+    xcb_generic_error_t *e = NULL;
+
+#define libxcb_error() \
+    (e = xcb->xcb_request_check(conn, vc), e != NULL)
+
+    /* Initialize the XInput2 externsion */
+    if (X11_check_xinput2_extension(conn, xcb))
+        goto_error("The XInput2 extension is not available");
+
+    i->xinput_ext_data = xcb->xcb_get_extension_data(conn,
+        xcb->xinput.xcb_input_id);
+    if (i->xinput_ext_data == NULL)
+        goto_error("Couldn't get XInput extension data");
+
+    /* Get the master keyboard and master mouse device IDs */
+    if (get_master_input_devices(
+            &i->master_mouse_id, &i->master_keyboard_id, conn, xcb
+    )) {
+        goto_error("Failed to query master input device IDs");
+    }
+
+    /* Select the events we want to receive from Xi2 */
+
+    /* Since libxcb has an incorrect struct definition,
+     * we have to make our own lol.
+     * This is taken directly from the Xi2 protocol specification
+     * (https://www.x.org/archive/X11R7.5/doc/inputproto/XI2proto.txt) */
+    const struct xi2_event_mask {
+        xcb_input_device_id_t deviceid;
+        u16 mask_len; /* Length of the mask in 4-byte units */
+        xcb_input_xi_event_mask_t mask;
+    } keyboard_mask = {
+        .deviceid = i->master_keyboard_id,
+        .mask_len = 1,
+        .mask = XCB_INPUT_XI_EVENT_MASK_KEY_PRESS
+            | XCB_INPUT_XI_EVENT_MASK_KEY_RELEASE,
+    }, mouse_mask = {
+        .deviceid = i->master_mouse_id,
+        .mask_len = 1,
+        .mask = XCB_INPUT_XI_EVENT_MASK_MOTION
+            | XCB_INPUT_XI_EVENT_MASK_BUTTON_PRESS
+            | XCB_INPUT_XI_EVENT_MASK_BUTTON_RELEASE,
+    };
+
+    vc = xcb->xinput.xcb_input_xi_select_events_checked(conn, win,
+        1, (const xcb_input_event_mask_t *)&keyboard_mask);
+    if (libxcb_error())
+        goto_error("Failed to enable keyboard input handling with Xi2");
+
+    vc = xcb->xinput.xcb_input_xi_select_events_checked(conn, win,
+        1, (const xcb_input_event_mask_t *)&mouse_mask);
+    if (libxcb_error())
+        goto_error("Failed to enable mouse input handling with Xi2");
+
+    /* Allocate the key symbols struct, used by keyboard-x11
+     * to map keycodes received from events to keysyms */
+    i->key_symbols = xcb->xcb_key_symbols_alloc(conn);
+    if (i->key_symbols == NULL)
+        goto_error("Failed to allocate key symbols");
+
+    /* Initialize the interfaces to `p_keyboard` and `p_mouse` */
+    i->registered_keyboard = NULL;
+    i->keyboard_deregistration_ack = p_mt_cond_create();
+    atomic_store(&i->keyboard_deregistration_notify, false);
+
+    i->registered_mouse = NULL;
+    i->mouse_deregistration_ack = p_mt_cond_create();
+    atomic_store(&i->mouse_deregistration_notify, false);
+
+    return 0;
+
+err:
+    if (e != NULL) {
+        u_nfree(&e);
+    }
+    return 1;
+
+#undef libxcb_error
+}
+
 static i32 init_acceleration(struct window_x11 *win, enum p_window_flags flags)
 {
     /* Decide which acceleration modes to try
@@ -696,14 +841,14 @@ static void render_destroy_egl(struct x11_render_egl_ctx *egl_rctx,
     egl_rctx->initialized_ = false;
 }
 
-static i32 intern_atom(struct window_x11 *win,
-    const char *atom_name, xcb_atom_t *o)
+static i32 intern_atom(const char *atom_name, xcb_atom_t *o,
+    xcb_connection_t *conn, const struct libxcb *xcb)
 {
-    xcb_intern_atom_cookie_t cookie = win->xcb.xcb_intern_atom(win->conn,
+    xcb_intern_atom_cookie_t cookie = xcb->xcb_intern_atom(conn,
         false, strlen(atom_name), atom_name);
 
     xcb_generic_error_t *err = NULL;
-    xcb_intern_atom_reply_t *reply = win->xcb.xcb_intern_atom_reply(win->conn,
+    xcb_intern_atom_reply_t *reply = xcb->xcb_intern_atom_reply(conn,
         cookie, &err);
 
     if (err) {
@@ -737,23 +882,21 @@ static bool query_extension(const char *name,
 }
 
 static i32 get_master_input_devices(
-    struct window_x11 *win,
     xcb_input_device_id_t *master_mouse_id,
-    xcb_input_device_id_t *master_keyboard_id
+    xcb_input_device_id_t *master_keyboard_id,
+    xcb_connection_t *conn, const struct libxcb *xcb
 )
 {
     xcb_input_xi_query_device_cookie_t cookie =
-        win->xcb.xinput.xcb_input_xi_query_device(win->conn,
-            XCB_INPUT_DEVICE_ALL);
+        xcb->xinput.xcb_input_xi_query_device(conn, XCB_INPUT_DEVICE_ALL);
 
     xcb_input_xi_query_device_reply_t *reply =
-        win->xcb.xinput.xcb_input_xi_query_device_reply(win->conn,
-            cookie, NULL);
+        xcb->xinput.xcb_input_xi_query_device_reply(conn, cookie, NULL);
     if (reply == NULL)
         return -1;
 
     xcb_input_xi_device_info_iterator_t iterator =
-        win->xcb.xinput.xcb_input_xi_query_device_infos_iterator(reply);
+        xcb->xinput.xcb_input_xi_query_device_infos_iterator(reply);
 
     bool found_keyboard = false, found_mouse = false;
     while (iterator.rem > 0 && !(found_keyboard && found_mouse)) {
@@ -767,24 +910,25 @@ static i32 get_master_input_devices(
             found_mouse = true;
         }
 
-        win->xcb.xinput.xcb_input_xi_device_info_next(&iterator);
+        xcb->xinput.xcb_input_xi_device_info_next(&iterator);
     }
 
     u_nfree(&reply);
     return found_mouse && found_keyboard ? 0 : 1;
 }
 
-static void send_dummy_event_to_self(struct window_x11 *win)
+static void send_dummy_event_to_self(xcb_window_t win,
+    xcb_connection_t *conn, const struct libxcb *xcb)
 {
     /* XCB will always copy 32 bytes from `ev` */
     char ev_base[32] = { 0 };
     xcb_expose_event_t *ev = (xcb_expose_event_t *)ev_base;
 
     ev->response_type = XCB_EXPOSE;
-    ev->window = win->win;
+    ev->window = win;
 
-    (void) win->xcb.xcb_send_event(win->conn, false, win->win,
+    (void) xcb->xcb_send_event(conn, false, win,
             XCB_EVENT_MASK_EXPOSURE, ev_base);
-    if (win->xcb.xcb_flush(win->conn) <= 0)
+    if (xcb->xcb_flush(conn) <= 0)
         s_log_error("xcb_flush failed!");
 }